提交分页和返回类封装

+++ follow-provider-user-8002/src/main/java/com/follow/controller/UserController.java
@@ -45,7 +45,7 @@ public class UserController {
 
     public User get_Message(@PathVariable("id") String id){
 
-        return new User().setId(Long.valueOf(id)).setUsername("没有这个用户").setPassword("密码不能告诉你");
+        return new User().setId(Integer.valueOf(id)).setUsername("没有这个用户").setPassword("密码不能告诉你");
     }
